MI5-114 Razdvojeni je ProcessingItem od PreviousProcessingsView
